CopilotAHK/requirements.txt
pyautogui>=0.9.54
Pillow>=10.0.0 
gitingest
psutil>=5.9.0
pytest 
pyperclip
watchdog>=3.0.0
plyer>=2.1
pystray>=0.19.5 